+++ modules/home-manager/programs/default.nix
@@ -4,7 +4,6 @@
     ./fastfetch.nix
     ./firefox.nix
     ./ghostty.nix
-    ./higgs-boson-firefox.nix
     ./gpg.nix
     ./kitty.nix
     ./neovim.nix