+++ modules/nixos/desktop/default.nix
@@ -3,8 +3,6 @@
     ./awesomewm.nix
     ./xfce.nix
     ./newm.nix
-    ./sddm.nix
-    ./ly.nix
     ./lightdm.nix
   ];
 }