+++ modules/nixos/security/default.nix
@@ -2,6 +2,7 @@
   imports = [
     ./doas.nix
     ./fail2ban.nix
+    ./firewall.nix
     ./packages.nix
     ./pam.nix
     ./ssh.nix